parser/requirements.txt
paddleocr
paddlepaddle
pydantic
tqdm
camelot-py[cv]